#147-transfer ui items into dedicated package
Intermediate commit. Number of reps per series completed and tested.
